+++ srcs/main.c
@@ -6,26 +6,12 @@
 /*   By: max <max@student.42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/27 07:35:54 by max               #+#    #+#             */
-/*   Updated: 2024/09/27 21:57:28 by max              ###   ########.fr       */
+/*   Updated: 2024/09/28 06:23:07 by max              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-long int get_precise_timestamp_in_ms(void)
-{
-    struct timeval tv;
-
-    // Boucle jusqu'à ce que la partie microsecondes soit proche de 0
-    while (1)
-    {
-        gettimeofday(&tv, NULL);
-        if (tv.tv_usec < 10) // Ici, on vérifie si on est proche du début d'une nouvelle milliseconde
-            break;
-    }
-
-    return ((tv.tv_sec * 1000) + (tv.tv_usec / 1000));
-}
 
 int update_limit(t_main_data *main_data, int i)
 {
@@ -35,33 +21,6 @@ int update_limit(t_main_data *main_data, int i)
 	reached_meals_limit = main_data->philosophers[i].meals_limit_reached;
 	pthread_mutex_unlock(&main_data->shared_data.meals_limit[i]);
 	return (reached_meals_limit);
-}
-
-bool check_death_and_meals_limit(t_main_data *main_data)
-{
-	int meals_completed;
-
-	meals_completed = update_limit_meals(main_data);
-
-	if (main_data->any_dead)
-		return (true);
-	if (main_data->has_meal_limit && meals_completed == main_data->shared_data.args.number_of_philosophers)
-	{
-		print_simulation_stop(main_data);
-		return (true);
-	}
-	return (false);
-}
-
-bool create_thread_array(t_main_data *main_data)
-{
-	main_data->threads = malloc(sizeof(pthread_t) * main_data->shared_data.args.number_of_philosophers);
-	if (main_data->threads == NULL)
-	{
-		print_error("Malloc failed");
-		return (false);
-	}
-	return (true);
 }
 
 int main(int argc, char **argv)
