+++ srcs/utils.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   utils.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mmauchre <mmauchre@student.42.fr>          +#+  +:+       +#+        */
+/*   By: max <max@student.42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/31 12:53:26 by max               #+#    #+#             */
-/*   Updated: 2024/09/21 15:50:11 by mmauchre         ###   ########.fr       */
+/*   Updated: 2024/09/28 06:22:58 by max              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,13 +24,6 @@ int	ft_strlen(char *str)
 	return (i);
 }
 
-long int	get_timestamp_in_ms(void)
-{
-	struct timeval	tv;
-
-	gettimeofday(&tv, NULL);
-	return ((tv.tv_sec * 1000) + (tv.tv_usec / 1000));
-}
 
 long int	ft_atoi(char *str)
 {
@@ -67,4 +60,20 @@ void	update_last_eaten_timestamp(t_philosopher *philosopher)
 	pthread_mutex_lock(&philosopher->shared_data->time_last_meal[i]);
 	philosopher->last_eaten_timestamp = get_timestamp_in_ms();
 	pthread_mutex_unlock(&philosopher->shared_data->time_last_meal[i]);
+}
+
+bool check_death_and_meals_limit(t_main_data *main_data)
+{
+	int meals_completed;
+
+	meals_completed = update_limit_meals(main_data);
+
+	if (main_data->any_dead)
+		return (true);
+	if (main_data->has_meal_limit && meals_completed == main_data->shared_data.args.number_of_philosophers)
+	{
+		print_simulation_stop(main_data);
+		return (true);
+	}
+	return (false);
 }
